requirements.txt
asgiref==3.8.1
Django==5.2.3
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
PyJWT==2.9.0
PyYAML==6.0.2
psycopg2-binary==2.9.10
GDAL==3.8.4
django-mptt==0.17.0
django-filter==25.1
Pillow==11.3.0




